+++ init/indexer.js
@@ -2,30 +2,16 @@ const Promise = require('bluebird');
 const request = require('request-promise');
 const retry = require('p-retry');
 const { unraw } = require('unraw');
-// const cloudscraper = require('cloudscraper').defaults({
-//   agentOptions: {
-//     ciphers: [
-//       'TLS_AES_128_GCM_SHA256',
-//       'TLS_CHACHA20_POLY1305_SHA256',
-//       'TLS_AES_256_GCM_SHA384',
-//       'TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256'
-//     ].join(':')
-//   }
-// });
 const agentOptions = require('../utils/agent');
 const cloudscraper = require('cloudscraper').defaults({ agentOptions });
 const { db } = require('../utils/db');
 async function indexer () {
   await db.transaction(async trx => {
-    const postsData = await trx('booru_posts')
-      .select('user', 'service')
-      .orderBy('added', 'desc')
-      .limit(10000);
-    await Promise.mapSeries(postsData, async (post) => {
-      const indexExists = await trx('lookup')
-        .where({ id: post.user, service: post.service });
-      if (indexExists.length) return;
-
+    const postsData = await trx.select('user', 'service')
+      .from({ post: 'booru_posts' })
+      .whereNotExists(db.select().from('lookup').whereRaw('id = post.user'))
+      .groupBy('user', 'service')
+    await Promise.map(postsData, async (post) => {
       switch (post.service) {
         case 'patreon': {
           const api = 'https://www.patreon.com/api/user';